requirements.txt
streamlit
torch
torchvision
transformers
pillow
numpy
pydicom
timm
pytorch-grad-cam
plotly
pandas
sqlalchemy
scikit-learn
reportlab
matplotlib
